requirements.txt
flask
requests
pytest
google-generativeai